libril: Move vendor commands header
Devices can now override this header in the device tree.
